+++ src/pdf.js
@@ -30,7 +30,6 @@ import {
   OPS,
   PasswordResponses,
   PermissionFlag,
-  removeNullCharacters,
   shadow,
   UnexpectedResponseException,
   UNSUPPORTED_FEATURES,
@@ -129,7 +128,6 @@ export {
   PDFWorker,
   PermissionFlag,
   PixelsPerInch,
-  removeNullCharacters,
   RenderingCancelledException,
   renderTextLayer,
   shadow,
