+++ src/shared/util.js
@@ -575,23 +575,6 @@ class AbortException extends BaseException {
   }
 }
 
-const NullCharactersRegExp = /\x00+/g;
-const InvisibleCharactersRegExp = /[\x01-\x1F]/g;
-
-/**
- * @param {string} str
- */
-function removeNullCharacters(str, replaceInvisible = false) {
-  if (typeof str !== "string") {
-    warn("The argument for removeNullCharacters must be a string.");
-    return str;
-  }
-  if (replaceInvisible) {
-    str = str.replace(InvisibleCharactersRegExp, " ");
-  }
-  return str.replace(NullCharactersRegExp, "");
-}
-
 function bytesToString(bytes) {
   assert(
     bytes !== null && typeof bytes === "object" && bytes.length !== undefined,
@@ -1185,7 +1168,6 @@ export {
   PasswordException,
   PasswordResponses,
   PermissionFlag,
-  removeNullCharacters,
   RenderingIntentFlag,
   setVerbosityLevel,
   shadow,
