+++ test/unit/ui_utils_spec.js
@@ -21,6 +21,7 @@ import {
   isPortraitOrientation,
   isValidRotation,
   parseQueryString,
+  removeNullCharacters,
 } from "../../web/ui_utils.js";
 
 describe("ui_utils", function () {
@@ -136,6 +137,30 @@ describe("ui_utils", function () {
       expect(parameters.size).toEqual(2);
       expect(parameters.get("key1")).toEqual("Value1");
       expect(parameters.get("key2")).toEqual("Value2");
+    });
+  });
+
+  describe("removeNullCharacters", function () {
+    it("should not modify string without null characters", function () {
+      const str = "string without null chars";
+      expect(removeNullCharacters(str)).toEqual("string without null chars");
+    });
+
+    it("should modify string with null characters", function () {
+      const str = "string\x00With\x00Null\x00Chars";
+      expect(removeNullCharacters(str)).toEqual("stringWithNullChars");
+    });
+
+    it("should modify string with non-displayable characters", function () {
+      const str = Array.from(Array(32).keys())
+        .map(x => String.fromCharCode(x) + "a")
+        .join("");
+      // \x00 is replaced by an empty string.
+      const expected =
+        "a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a";
+      expect(removeNullCharacters(str, /* replaceInvisible */ true)).toEqual(
+        expected
+      );
     });
   });
