+++ web/ui_utils.js
@@ -200,6 +200,24 @@ function parseQueryString(query) {
   return params;
 }
 
+const NullCharactersRegExp = /\x00/g;
+const InvisibleCharactersRegExp = /[\x01-\x1F]/g;
+
+/**
+ * @param {string} str
+ * @param {boolean} [replaceInvisible]
+ */
+function removeNullCharacters(str, replaceInvisible = false) {
+  if (typeof str !== "string") {
+    console.error(`The argument must be a string.`);
+    return str;
+  }
+  if (replaceInvisible) {
+    str = str.replace(InvisibleCharactersRegExp, " ");
+  }
+  return str.replace(NullCharactersRegExp, "");
+}
+
 /**
  * Use binary search to find the index of the first item in a given array which
  * passes a given condition. The items are expected to be sorted in the sense
@@ -838,6 +856,7 @@ export {
   parseQueryString,
   PresentationModeState,
   ProgressBar,
+  removeNullCharacters,
   RendererType,
   RenderingStates,
   roundToDivide,
